refactor: SHA256으로 계층 간소화
